escmd/requirements.txt
elasticsearch==7.17.12
PyYAML==6.0.1
Requests==2.31.0
rich==13.7.1
setuptools==53.0.0
urllib3==1.26.16
